+++ pages/main.tsx
@@ -1,4 +1,4 @@
-import { LoginLayout } from '@/components/common/layouts';
+import { LoginLayout } from '@/components/common';
 import { MainPage } from '@/components/main';
 import { ReactElement } from 'react';
